requirements.txt
numpy
scipy
opencv-python
opencv-contrib-python
tqdm
pyyaml
flake8
jupyter
tensorflow-cpu